# Conflicts:
#	radio/src/gui/212x64/widgets.cpp

+++ radio/src/strhelpers.cpp
@@ -49,7 +49,7 @@ int8_t char2idx(char c)
 {
   if (c == '_') return 37;
 #if LEN_SPECIAL_CHARS > 0
-  if (c < 0 && c+128 <= LEN_SPECIAL_CHARS) return 41 + (c+128);
+  if ((int8_t)c < 0 && c+128 <= LEN_SPECIAL_CHARS) return 41 + (c+128);
 #endif
   if (c >= 'a') return 'a' - c - 1;
   if (c >= 'A') return c - 'A' + 1;
